Refractor the home module, move it 1-level up at hypha.home
