+++ hypha/apply/funds/management/commands/seed_community_lab_application.py
@@ -6,9 +6,9 @@ from django.db import transaction
 
 from hypha.apply.funds.models import ApplicationForm, LabType
 from hypha.apply.funds.models.forms import LabBaseForm, LabBaseReviewForm
-from hypha.apply.home.models import ApplyHomePage
 from hypha.apply.review.models import ReviewForm
 from hypha.apply.users.groups import STAFF_GROUP_NAME
+from hypha.home.models import ApplyHomePage
 
 CL_FUND_TITLE = "Community lab (archive fund)"
