+++ hypha/apply/funds/management/commands/seed_concept_note.py
@@ -11,9 +11,9 @@ from hypha.apply.funds.models.forms import (
     ApplicationBaseForm,
     ApplicationBaseReviewForm,
 )
-from hypha.apply.home.models import ApplyHomePage
 from hypha.apply.review.models import ReviewForm
 from hypha.apply.users.groups import STAFF_GROUP_NAME
+from hypha.home.models import ApplyHomePage
 
 CN_ROUND_TITLE = "Internet Freedom Fund (archive round)"
 CN_FUND_TITLE = "Internet Freedom Fund (archive fund)"
