+++ hypha/apply/funds/management/commands/send_reminders.py
@@ -6,7 +6,7 @@ from django.utils import timezone
 
 from hypha.apply.activity.messaging import messenger
 from hypha.apply.funds.models import Reminder
-from hypha.apply.home.models import ApplyHomePage
+from hypha.home.models import ApplyHomePage
 
 
 class Command(BaseCommand):
